Add syncing README.md to Docker hub.

+++ scripts/build-sfepy-docker.sh
@@ -4,13 +4,18 @@ RED='\033[0;31m'
 NC='\033[0m'
 
 function usage {
-    echo -e "Usage: $0 [-v version_tag] [-r repository] [-d build_dirs] [-p] \n"
+    echo -e "Usage: $0 [-v version_tag] [-r repository] [-d build_dirs] [-p] [-s] \n"
     exit 1
 }
 
 REPO="sfepy"
 VERSION=""
+
 PUSH="NO"
+
+SYNC="NO"
+README="README.md"
+SYNC_TOKEN_FILE=".app-access-token-dockerhub"
 BUILD_DIRS="sfepy-desktop"
 
 while [[ $# -gt 0 ]]
@@ -37,7 +42,11 @@ case $key in
     PUSH="YES"
     shift
     ;;
-  *)
+   -s|--sync-readme)
+    SYNC="YES"
+    shift
+    ;;
+ *)
     usage
     exit 1
     ;;
@@ -51,24 +60,36 @@ if [[ -z "$VERSION" ]]; then
           sed 's/.*\///; s/\^{}//; s/release_//')
 fi
 
-CWD="$(cd -P -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd -P)"
-cd "$CWD"/.. || exit
+SCRIPT_BIN="$(cd -P -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd -P)"
+cd "$SCRIPT_BIN"/.. || exit
 
 for dir in ${BUILD_DIRS}
 do
   cd "${dir}" || exit
-
-  echo -e "${RED}Building Docker image(s): ${dir}:${VERSION}${NC}"
-  docker build --rm .  --build-arg SFEPY_RELEASE="${VERSION}" \
-         -t "$REPO/${dir}" -t "$REPO/${dir}:$VERSION"
-  echo -e "Done.\n"
 
   if [[ "$PUSH" = "YES" ]]; then
     echo -e "${RED}Pushing Docker images to $REPO repository${NC}"
     docker push "$REPO/${dir}"
     docker push "$REPO/${dir}:$VERSION"
     echo -e "Done.\n"
+    exit 0
   fi
+
+  if [[ "$SYNC" = "YES" ]]; then
+    echo -e "${RED}Syncing ${dir}/${README} to Docker description file...${NC}"
+    "${SCRIPT_BIN}"/docker-sync-readme.py \
+        --username="kejzlar" \
+        --readme="${README}" \
+        --repo="$REPO/${dir}" \
+        --password="$(<"${SCRIPT_BIN}"/${SYNC_TOKEN_FILE})"
+    echo -e "Done.\n"
+    exit 0
+  fi
+
+  echo -e "${RED}Building Docker image(s): ${dir}:${VERSION}${NC}"
+  docker build --rm .  --build-arg SFEPY_RELEASE="${VERSION}" \
+         -t "$REPO/${dir}" -t "$REPO/${dir}:$VERSION"
+  echo -e "Done.\n"
 
   cd ..
 done
